Conflicts:
	Repository/BaseRepository/Model.Views.cs
	Repository/BaseRepository/Model.edmx

+++ Models/DomainModels/Widget.cs
@@ -8,7 +8,7 @@ namespace MPC.Models.DomainModels
 {
     public class Widget
     {
-        public int WidgetId { get; set; }
+        public long WidgetId { get; set; }
         public string WidgetCode { get; set; }
         public string WidgetName { get; set; }
         public string WidgetControlName { get; set; }
